pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "pons",
  "tools/netedit",
  "tools/ponsanim",
  "tools/plot_contacts",
  "tools/scenariorunner",
]

[project]
name = "pons-dtn"
version = "0.2.1"
authors = [
  { name = "Lars Baumgaertner", email = "1264131+gh0st42@users.noreply.github.com" },
]
description = "The Python Opportunistic Network Simulator (PONS) is a discrete-event simulator for opportunistic/DTN networks."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Communications",
]
keywords = [
  "opportunistic networks",
  "discrete-event simulation",
  "DTN",
  "network simulation",
  "networking",
  "simulator",
  "delay-tolerant networks",
  "disruption-tolerant networks",
]
dependencies = [
  "networkx~=3.3",
  "simpy~=4.1.1",
  "pillow~=10.4.0",
  "python-dateutil~=2.9.0.post0",
  "pandas~=2.2.3",
  "matplotlib~=3.10.1",
]

[project.optional-dependencies]
mp4 = ["imageio[ffmpeg]"]

[project.scripts]
ponsanim = "ponsanim.ponsanim:main"
plot-contacts = "plot_contacts.plot_contacts:main"
scenariorunner = "scenariorunner.scenariorunner:main"

[project.gui-scripts]
netedit = "netedit.netedit:main"

[project.urls]
Homepage = "https://github.com/gh0st42/pons"
Issues = "https://github.com/gh0st42/pons/issues"
Repository = "https://github.com/gh0st42/pons.git"
